+++ CHANGELOG.md
@@ -40,6 +40,7 @@ develop
 - Fix an issue where text view drawing was distorted while resizing window.
 - Fix an issue where line endings of a document that has a line ending chacter at the beginning of the file cannot be interpreted its line ending type correctly.
 - Fix an issue where character inspector returned always `U+000A` (LF) for line ending even the actual line ending of the document is not LF.
+- Fix character count with a single regional indicator symbol.
 - Fix wrong undo action name on encoding conversion via script.